refactor(challenge): ChallengeView/PortfolioView 중복 파생로직을 공유 hook으로 추출 (T3)
두 뷰에 완전히 동일하게 중복돼 있던 head(isResumeTemplate·receivedContent·
weekText·hasReviews·challengeTransformed·styles + active/faq 쿼리 + init 이펙트)를
useChallengeDetailView 훅으로 통합.
- ChallengeView 357→257, ChallengePortfolioView 405→302줄
- 미사용 import/enum destructure 정리
동작 변경 없음. tsc 0 / lint 통과.
